factor out components
