+++ garden/src/features/modal/components/progress/OverallProgress.tsx
@@ -1,0 +1,75 @@
+import { CheckCircle2 } from "lucide-react";
+import { cn } from "@/utils/form.utils";
+
+interface OverallProgressProps {
+  currentPhase: number;
+  isCompleted?: boolean;
+}
+
+export const OverallProgress = ({ currentPhase, isCompleted = false }: OverallProgressProps) => {
+  const phases = [
+    { number: 1, label: "Upload Modal App" },
+    { number: 2, label: "Configure Garden" }
+  ];
+
+  // Calculate progress percentage - stop just before next point until completed
+  const getProgressPercentage = () => {
+    if (currentPhase <= 1 && !isCompleted) return 85; // Stop just before first point completes
+    if (currentPhase === 1 && isCompleted) return 100; // Phase 1 complete
+    if (currentPhase >= 2) return 100; // Phase 2 or beyond
+    return 0; // Shouldn't reach here
+  };
+
+  return (
+    <div className="mb-6">
+      <div className="mb-2">
+        <h3 className="text-base font-medium text-gray-700">Publishing Progress</h3>
+      </div>
+      <div className="relative max-w-3xl mx-auto">
+        {/* Progress Bar Background */}
+        <div className="absolute left-0 top-5 h-1 w-full rounded-full bg-gray-200"></div>
+        
+        {/* Progress Bar Fill - This shows completed steps */}
+        <div 
+          className="absolute left-0 top-5 h-1 rounded-full bg-primary transition-all duration-300 ease-in-out"
+          style={{ width: `${getProgressPercentage()}%` }}
+        ></div>
+        
+        {/* Phases - Using a max-width container with better spacing */}
+        <div className="relative max-w-md mx-auto px-8">
+          <div className="flex w-full justify-between">
+            {phases.map((phase) => (
+              <div 
+                key={phase.number} 
+                className="flex flex-col items-center"
+              >
+                <div 
+                  className={cn(
+                    "relative z-10 flex h-10 w-10 items-center justify-center rounded-full border-2 transition-colors duration-300",
+                    currentPhase >= phase.number 
+                      ? "border-primary bg-primary text-white"
+                      : "border-gray-300 bg-white text-gray-400"
+                  )}
+                >
+                  {currentPhase > phase.number ? (
+                    <CheckCircle2 className="h-7 w-7" />
+                  ) : (
+                    <span className="text-base font-semibold">{phase.number}</span>
+                  )}
+                </div>
+                <span 
+                  className={cn(
+                    "mt-2 text-sm font-medium",
+                    currentPhase >= phase.number ? "text-gray-900" : "text-gray-500"
+                  )}
+                >
+                  {phase.label}
+                </span>
+              </div>
+            ))}
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+};
